MultiplayerAirHockey - Update commit - Hour 5
Complicated code with some mathematics involved. Made a separate html to test the air hockey game. There should be a grey box with a 8:5 ration and a green start game button. The air hockey table should be an 8:5 ratio dimension. It should track the speed and direction of the paddle. And on the exact moment the paddle touches the puck, the puck will obtain the speed and direction value of the paddle (only at that instant) then travel by momentum (it loses 1 speed value every 0.5 seconds). The puck should bounce off the wall at the exact same speed, but different direction (using reflection law where angle of incident ray = angle of reflection ray).