MultiplayerAirHockey - Update commit - Hour 7
Spent most of the time debugging😭 but i managed to implement the air hockey game code into the main server code. The Hockey game window should be hidden when not joining any room, and it will appear when you hosted or joined a room code. The start button will only turn green and be clickable when more than 2 players are in the room