MultiplayerAirHockey - Update commit - Hour 8
Made the multiplayer system. The host can only control the left side paddle, and the joiner can only control the right side paddle. Physics calculation of the puck is done on the host client. The position of the left side paddle should mirror to the joiner client, and the position of the right paddle should mirror to the host client